requirements.txt
Flask==2.0.2
Werkzeug==2.2.2
Django
whitenoise
wsgi-basic-auth
tweepy
google-auth
google-api-python-client
google-auth-oauthlib
